requirements.txt
loguru~=0.6.0
requests~=2.28.1
Pillow~=9.3.0
python-dotenv~=0.21.0
selenium~=4.6.0
PyJWT~=2.6.0
APScheduler==3.9.1
